requirements.txt
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
elasticsearch==7.5.1
idna==2.8
python-dotenv==0.12.0
requests==2.22.0
six==1.14.0
urllib3==1.25.8
validators==0.14.2
